feat(DTFS2-6745): update e2e filestructure to commonise user tasks
