+++ e2e-tests/gef/cypress.config.js
@@ -1,6 +1,5 @@
 const { defineConfig } = require('cypress');
-const { createTasks } = require('./cypress/support/tasks');
-const db = require('../support/db-client');
+const { createUserTasks } = require('../support/user-tasks');
 
 module.exports = defineConfig({
   dealApiProtocol: 'http://',
@@ -28,12 +27,7 @@ module.exports = defineConfig({
     specPattern: 'cypress/e2e/**/*.spec.js',
     setupNodeEvents(on, config) {
       const { dbName, dbConnectionString } = config;
-      const connectionOptions = { dbName, dbConnectionString };
-      const usersCollectionName = 'users';
-
-      const getUsersCollection = () => db.getCollection(usersCollectionName, connectionOptions);
-
-      on('task', createTasks(getUsersCollection)); // TODO dtfs2-6745: consider pull out createtasks to a global area
+      on('task', createUserTasks({ dbName, dbConnectionString })); // TODO dtfs2-6745: consider pull out createtasks to a global area
     },
   },
   experimentalCspAllowList: ['child-src', 'default-src', 'frame-src', 'form-action', 'script-src', 'script-src-elem'],
