+++ e2e-tests/gef/cypress/support/commands/portal/login.js
@@ -2,13 +2,13 @@ const relative = require('../../relativeURL');
 const signInLink = require('../../../e2e/pages/login/sign-in-link');
 
 const login = ({ username, password }) => {
-  cy.resetUserStatusAndNumberOfSignInLinks(username);
+  cy.resetPortalUserStatusAndNumberOfSignInLinks(username);
   cy.enterUsernameAndPassword({ username, password });
 
   cy.url().should('eq', relative('/login/check-your-email'));
 
   const signInToken = '6569ca7a6fd828f925e07c6e';
-  cy.overrideUserSignInTokenByUsername({ username, newSignInToken: signInToken });
+  cy.overridePortalUserSignInTokenByUsername({ username, newSignInToken: signInToken });
   cy.getUserByUsername(username).then(({ _id }) => {
     signInLink.visit({ token: signInToken, userId: _id });
   });
