+++ e2e-tests/gef/cypress/support/commands/portal/overridePortalUserSignInTokenByUsername.js
@@ -1,0 +1,1 @@
+module.exports = ({ username, newSignInToken }) => cy.task('overridePortalUserSignInTokenByUsername', { username, newSignInToken });
