--- e2e-tests/gef/cypress/support/commands/portal/overrideUserSignInTokenByUsername.js
@@ -1,1 +1,0 @@
-module.exports = ({ username, newSignInToken }) => cy.task('overrideUserSignInTokenByUsername', { username, newSignInToken });
