+++ e2e-tests/gef/cypress/support/commands/portal/resetPortalUserStatusAndNumberOfSignInLinks.js
@@ -1,0 +1,1 @@
+module.exports = (username) => cy.task('resetPortalUserStatusAndNumberOfSignInLinks', username);
