--- e2e-tests/gef/cypress/support/commands/portal/resetUserStatusAndNumberOfSignInLinks.js
@@ -1,1 +1,0 @@
-module.exports = (username) => cy.task('resetUserStatusAndNumberOfSignInLinks', username);
