+++ e2e-tests/portal/cypress/support/commands.js
@@ -71,10 +71,10 @@ Cypress.Commands.add('enterUsernameAndPassword', require('./portal/enterUsername
 Cypress.Commands.add('getUserByUsername', require('./portal/getUserByUsername'));
 Cypress.Commands.add('login', require('./portal/login'));
 Cypress.Commands.add('loginGoToDealPage', require('./portal/loginGoToDealPage'));
-Cypress.Commands.add('overrideUserSignInTokenByUsername', require('./portal/overrideUserSignInTokenByUsername'));
+Cypress.Commands.add('overridePortalUserSignInTokenByUsername', require('./portal/overridePortalUserSignInTokenByUsername'));
 Cypress.Commands.add('passRedLine', require('./portal/passRedLine'));
 Cypress.Commands.add('renameDeal', require('./portal/renameDeal'));
-Cypress.Commands.add('resetUserStatusAndNumberOfSignInLinks', require('./portal/resetUserStatusAndNumberOfSignInLinks'));
+Cypress.Commands.add('resetPortalUserStatusAndNumberOfSignInLinks', require('./portal/resetPortalUserStatusAndNumberOfSignInLinks'));
 Cypress.Commands.add('userSetPassword', require('./portal/userSetPassword'));
 
 // commands that add/edit facilities directly in central API
