+++ e2e-tests/portal/cypress/support/portal/overridePortalUserSignInTokenByUsername.js
@@ -1,0 +1,1 @@
+module.exports = ({ username, newSignInToken }) => cy.task('overridePortalUserSignInTokenByUsername', { username, newSignInToken });
