--- e2e-tests/portal/cypress/support/portal/overrideUserSignInTokenByUsername.js
@@ -1,1 +1,0 @@
-module.exports = ({ username, newSignInToken }) => cy.task('overrideUserSignInTokenByUsername', { username, newSignInToken });
