+++ e2e-tests/portal/cypress/support/portal/resetPortalUserStatusAndNumberOfSignInLinks.js
@@ -1,0 +1,1 @@
+module.exports = (username) => cy.task('resetPortalUserStatusAndNumberOfSignInLinks', username);
