--- e2e-tests/portal/cypress/support/portal/resetUserStatusAndNumberOfSignInLinks.js
@@ -1,1 +1,0 @@
-module.exports = (username) => cy.task('resetUserStatusAndNumberOfSignInLinks', username);
